+++ catalyst/contrib/data/cv/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
 
 from .mixins import *
+from .reader import *
 from .transforms import *
