+++ catalyst/contrib/utils/__init__.py
@@ -6,6 +6,8 @@ import os
 
 logger = logging.getLogger(__name__)
 
+from catalyst.tools import settings
+
 from .argparse import boolean_flag
 from .compression import pack, pack_if_needed, unpack, unpack_if_needed
 from .confusion_matrix import (
@@ -14,17 +16,27 @@ from .confusion_matrix import (
     calculate_confusion_matrix_from_tensors,
 )
 from .dataset import create_dataset, split_dataset_train_test, create_dataframe
-from .image import (
-    has_image_extension,
-    imread,
-    imwrite,
-    imsave,
-    mask_to_overlay_image,
-    mimread,
-    mimwrite_with_meta,
-    tensor_from_rgb_image,
-    tensor_to_ndimage,
-)
+
+try:
+    from .image import (
+        has_image_extension,
+        imread,
+        imwrite,
+        imsave,
+        mask_to_overlay_image,
+        mimread,
+        mimwrite_with_meta,
+        tensor_from_rgb_image,
+        tensor_to_ndimage,
+    )
+except ImportError as ex:
+    if settings.cv_required:
+        logger.warning(
+            "some of catalyst-cv dependencies not available,"
+            " to install dependencies, run `pip install catalyst[cv]`."
+        )
+        raise ex
+
 from .misc import (
     args_are_not_none,
     make_tuple,
@@ -48,13 +60,12 @@ from .pandas import (
 )
 from .parallel import parallel_imap, tqdm_parallel_imap, get_pool
 from .serialization import deserialize, serialize
-from catalyst.tools import settings
 
 try:
     import plotly  # noqa: F401
     from .plotly import plot_tensorboard_log, plot_metrics
 except ImportError as ex:
-    if os.environ.get("USE_PLOTLY", "0") == "1":
+    if settings.plotly_required:
         logger.warning(
             "plotly not available, to install plotly,"
             " run `pip install plotly`."
