+++ catalyst/data/scripts/image2embedding.py
@@ -9,8 +9,8 @@ from tqdm import tqdm
 import torch
 from torchvision import transforms
 
+from catalyst.contrib.data.cv import ImageReader
 from catalyst.contrib.models.cv import ResnetEncoder
-from catalyst.data import ImageReader
 from catalyst.dl import utils
 
 cv2.setNumThreads(0)
