+++ requirements/requirements-cv.txt
@@ -1,2 +1,5 @@
 albumentations==0.4.3
+imageio
+scikit-image>=0.14.2
 segmentation-models-pytorch==0.1.0
+torchvision>=0.3.0
