+++ requirements/requirements.txt
@@ -15,14 +15,8 @@ ipython
 tensorboard>=1.14.0
 tensorboardX
 crc32c>=1.7
+opencv-python-headless
 plotly>=4.1.0
-
-# cv -> catalyst[cv]
-torchvision>=0.3.0
-Pillow
-imageio
-opencv-python
-scikit-image>=0.14.2
 
 # Used in scripts  -> contrib (?)
 matplotlib
